+++ README.md
@@ -6,127 +6,39 @@
   |----README.md
   |----Project_report.docs  //NUFINISH
   |----Makefile             //NUFINISH
-  |----Biopartite.cpp
-  |----adj_list.cpp
-  |----adj_matrix.cpp
-  |----adj_matrix_pointer.cpp
-  |----adj_list_pointer.cpp
-  |----main_time_calculating.cpp
-  |----convert_list_to_matrix.cpp
-  |----list_add_neighbor_number.cpp
+  |----Bipartite.cpp
   |----test_cases
-     |----11.txt
-     |----12.txt
-     |----13.txt
-     |----test_adj_list.txt
-     |----test_adj_list_for_pointer.txt
-     |----test_adj_matrix.txt
-     |----benchmarks.jpg
+     |----test_cases.txt
+     |----test_cases.jpg
   |----open_banchmarks
-     |----1.txt
-     |----2.txt
-     |----3.txt
-     |----4.txt
-     |----5.txt
-     |----6.txt
-     |----7.txt
-     |----8.txt
-     |----9.txt
-     |----10.txt
-     |----test_adj_list.txt
-     |----test_adj_list_for_pointer.txt
-     |----test_adj_matrix.txt
+     |----open_banchmarks.txt
      |----open_benchmarks_1.jpg
      |----open_benchmarks_2.jpg
-  |----test_time
-     |----adj_list_time.cpp
-     |----adj_matrix_time.cpp
-     |----adj_matrix_pointer_time.cpp
-     |----adj_list_pointer_time.cpp
 ```
 ## input file:
 ```
-/test_cases/test_adj_matrix.txt
-for adj_matrix.cpp
-    adj_matrix_pointer.cpp
-    adj_matrix_time.cpp
-    adj_matrix_pointer_time.cpp
-
-/test_cases/test_adj_list.txt
-for adj_list.cpp
-    adj_list_time.cpp
-
-/test_cases/test_adj_list_for_pointer.txt
-for adj_list_pointer.cpp
-    adj_list_pointer_time.cpp
-
-/open_benchmarks/test_adj_matrix.txt
-for adj_matrix.cpp
-    adj_matrix_pointer.cpp
-    adj_matrix_test.cpp
-    adj_matrix_pointer_test.cpp
-    
-/open_benchmarks/test_adj_list.txt
-for adj_list.cpp
-    adj_list_time.cpp
-
-/open_benchmarks/test_adj_list_for_pointer.txt
-for adj_list_pointer.cpp
-    adj_list_pointer_time.cpp
-
-...
-```
-
-## test_case(test_adj_list.txt)
-```
-number of graphs
-number of vertex in this graph
-adj vertex of vertex(1)
-adj vertex of vertex(2)
-adj vertex of vertex(3)
-...
-
-number of vertex in this graph
-adj vertex of vertex(1)
-adj vertex of vertex(2)
-adj vertex of vertex(3)
+/test_cases/test_cases.txt
+/open_benchmarks/open_benchmarks.txt
 ...
 ```
 
 ## test_case(test_adj_list_for_pointer.txt)
 ```
-number of graphs
-number of vertex in this graph
-number_of_adj_vertexes adj_vertex_1_of_vertex(1)...
-number_of_adj_vertexes adj_vertex_1_of_vertex(2)...
-number_of_adj_vertexes adj_vertex_1_of_vertex(3)...
+number_of_graphs
+number_of_vertex_in_graph(1)
+number_of_adj_vertexes_of_vertex(1) adj_vertex_1_of_vertex(1) adj_vertex_2_of_vertex(1)...
+number_of_adj_vertexes_of_vertex(2) adj_vertex_1_of_vertex(2)...
+number_of_adj_vertexes_of_vertex(3) adj_vertex_1_of_vertex(3)...
 ...
 
-number of graphs
-number of vertex in this graph
-number_of_adj_vertexes adj_vertex_1_of_vertex(1)...
-number_of_adj_vertexes adj_vertex_1_of_vertex(2)...
-number_of_adj_vertexes adj_vertex_1_of_vertex(3)...
-...
-```
-
-## test_case(test_adj_matrix.txt)
-```
-number of graphs
-number of vertex in this graph
-adj matrix of this graph
-
-number of vertex in this graph
-adj matrix of this graph
+number_of_vertex_in_graph(2)
+number_of_adj_vertexes_of_vertex(1) adj_vertex_1_of_vertex(1) adj_vertex_2_of_vertex(1)...
+number_of_adj_vertexes_of_vertex(2) adj_vertex_1_of_vertex(2)...
+number_of_adj_vertexes_of_vertex(3) adj_vertex_1_of_vertex(3)...
 ...
 ```
 
-## test_time folder
-```
-精準到奈秒計算時間
-```
-
-## benchmarks
+## testcases
 ![image](test_cases/benchmarks.jpg)
 
 ## open benchmarks
