--- open_benchmarks/1.txt
@@ -1,15 +1,0 @@
-6
-7
-9
-9
-15
-1 11
-2 13
-12
-3 4 13
-14
-6
-8
-7 9
-10
-5
