--- open_benchmarks/10.txt
@@ -1,20 +1,0 @@
-13
-12
-14 15
-11
-14
-16 17
-19
-17 18
-10 20
-9 19
-4
-2
-1
-3 5
-3 16
-6 15
-6 8 20
-8
-7 10 20
-9 17 19
