--- open_benchmarks/2.txt
@@ -1,20 +1,0 @@
-13
-12
- 
-11
-14
-16
-19
-17 18
-20
-19
-4
-2
-1
-5
- 
-6
-8
-8
-7 10
-9
