--- open_benchmarks/3.txt
@@ -1,20 +1,0 @@
-13
-12
- 
-11 14
-14
-16
-19
-17 18
-14 20
-16 19
-4
-2
-1
-4 5 9
- 
-6 10
-8
-8
-7 10
-9
