--- open_benchmarks/4.txt
@@ -1,20 +1,0 @@
-2 10
-1 7
-4 7
-3 5
-4 6
-5 9
-2 3 8
-7 11
-6 12
-1 16
-8
-9 15
-14 16 17
-13 15
-12 14
-10 13
-13 18
-17 19
-18 20
-19
