--- open_benchmarks/5.txt
@@ -1,20 +1,0 @@
-11 13
-12
-13 14
-11 15
-14
-16
-19
-17 18
-20
-19
-1 4
-2
-1 3
-3 5 15
-4 14
-6
-8
-8
-7 10
-9
