--- open_benchmarks/6.txt
@@ -1,15 +1,0 @@
-6
-7 8
-9
-9
-15
-1 11
-2 13
-2 12
-3 4 13
-14
-6
-8 13
-7 9 12
-10
-5
