--- open_benchmarks/7.txt
@@ -1,15 +1,0 @@
-6 9
-7
-9
-9
-9 15
-1 11
-2 13
-11 12
-1 3 4 5 13
-14
-6 8
-8
-7 9
-10
-5
