--- open_benchmarks/8.txt
@@ -1,20 +1,0 @@
-2 3
-1 4
-1 5 6
-2 7
-3 8 9
-3
-4 10 11
-5
-5 12 13
-7
-7
-9 14
-9 15
-12 16 17
-13
-14 18 19
-14 20
-16
-16
-17
