--- open_benchmarks/9.txt
@@ -1,20 +1,0 @@
-13
-12
-14 15
-11
-7 14 16
-16 17
-5 17 19
-17 18 19
-10 20
-9 19
-4
-2
-1
-3 5
-3 16
-5 6 15
-6 7 8 20
-8
-7 8 10 20
-9 17 19
