--- open_benchmarks/test_adj_list.txt
@@ -1,205 +1,0 @@
-10
-15
-6
-7
-9
-9
-15
-1 11
-2 13
-12
-3 4 13
-14
-6
-8
-7 9
-10
-5
-
-20
-13
-12
- 
-11
-14
-16
-19
-17 18
-20
-19
-4
-2
-1
-5
- 
-6
-8
-8
-7 10
-9
-
-20
-13
-12
- 
-11 14
-14
-16
-19
-17 18
-14 20
-16 19
-4
-2
-1
-4 5 9
- 
-6 10
-8
-8
-7 10
-9
-
-20
-2 10
-1 7
-4 7
-3 5
-4 6
-5 9
-2 3 8
-7 11
-6 12
-1 16
-8
-9 15
-14 16 17
-13 15
-12 14
-10 13
-13 18
-17 19
-18 20
-19
-
-20
-11 13
-12
-13 14
-11 15
-14
-16
-19
-17 18
-20
-19
-1 4
-2
-1 3
-3 5 15
-4 14
-6
-8
-8
-7 10
-9
-
-15
-6
-7 8
-9
-9
-15
-1 11
-2 13
-2 12
-3 4 13
-14
-6
-8 13
-7 9 12
-10
-5
-
-15
-6 9
-7
-9
-9
-9 15
-1 11
-2 13
-11 12
-1 3 4 5 13
-14
-6 8
-8
-7 9
-10
-5
-
-20
-2 3
-1 4
-1 5 6
-2 7
-3 8 9
-3
-4 10 11
-5
-5 12 13
-7
-7
-9 14
-9 15
-12 16 17
-13
-14 18 19
-14 20
-16
-16
-17
-
-20
-13
-12
-14 15
-11
-7 14 16
-16 17
-5 17 19
-17 18 19
-10 20
-9 19
-4
-2
-1
-3 5
-3 16
-5 6 15
-6 7 8 20
-8
-7 8 10 20
-9 17 19
-
-20
-13
-12
-14 15
-11
-14
-16 17
-19
-17 18
-10 20
-9 19
-4
-2
-1
-3 5
-3 16
-6 15
-6 8 20
-8
-7 10 20
-9 17 19
