--- open_benchmarks/test_adj_matrix.txt
@@ -1,206 +1,0 @@
-10
-15
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 
-1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 
-0 0 1 1 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
-
-20
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
-
-20
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 
-0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 1 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
-
-20
-0 1 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
-1 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
-0 1 1 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 1 1 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 1 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-
-20
-0 0 0 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-1 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
-
-15
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 
-1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 
-0 1 0 0 0 0 0 0 0 0 0 1 0 0 0 
-0 0 1 1 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 1 0 0 
-0 0 0 0 0 0 1 0 1 0 0 1 0 0 0 
-0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
-
-15
-0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 1 
-1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 
-1 0 1 1 1 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
-
-20
-0 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 0 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 1 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 1 1 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 1 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 
-
-20
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 1 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 
-0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 
-0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 1 1 0 0 0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 1 0 1 0 
-
-20
-0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 
-0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 1 0 
-0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 1 
-0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 1 0 1 0 
-
