--- test_cases/11.txt
@@ -1,12 +1,0 @@
-7 9
-9
-7 12
-10
-9 12
-11
-1 3
- 
-1 2 5
-4
-6
-3 5
