--- test_cases/12.txt
@@ -1,12 +1,0 @@
-7 9
-9
-7 12
-10
-6 9
-5 11 12
-1 3
- 
-1 2 5
-4
-6
-3 6
