--- test_cases/13.txt
@@ -1,12 +1,0 @@
-7
-9
-7 12
-10
-9
-11 12
-1 3
- 
-2 5
-4
-6
-3 6
