--- test_cases/test_adj_list.txt
@@ -1,42 +1,0 @@
-3
-12
-7 9
-9
-7 12
-10
-9 12
-11
-1 3
- 
-1 2 5
-4
-6
-3 5
-
-12
-7 9
-9
-7 12
-10
-6 9
-5 11 12
-1 3
- 
-1 2 5
-4
-6
-3 6
-
-12
-7
-9
-7 12
-10
-9
-11 12
-1 3
- 
-2 5
-4
-6
-3 6
