--- test_cases/test_adj_matrix.txt
@@ -1,43 +1,0 @@
-3
-12
-0 0 0 0 0 0 1 0 1 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 1 0 0 1 
-0 0 0 0 0 0 0 0 0 0 1 0 
-1 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 
-1 1 0 0 1 0 0 0 0 0 0 0 
-0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 1 0 1 0 0 0 0 0 0 0 
-
-12
-0 0 0 0 0 0 1 0 1 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 1 0 0 1 0 0 0 
-0 0 0 0 1 0 0 0 0 0 1 1 
-1 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 
-1 1 0 0 1 0 0 0 0 0 0 0 
-0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 1 0 0 1 0 0 0 0 0 0 
-
-12
-0 0 0 0 0 0 1 0 0 0 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 
-0 0 0 0 0 0 1 0 0 0 0 1 
-0 0 0 0 0 0 0 0 0 1 0 0 
-0 0 0 0 0 0 0 0 1 0 0 0 
-0 0 0 0 0 0 0 0 0 0 1 1 
-1 0 1 0 0 0 0 0 0 0 0 0 
-0 0 0 0 0 0 0 0 0 0 0 0 
-0 1 0 0 1 0 0 0 0 0 0 0 
-0 0 0 1 0 0 0 0 0 0 0 0 
-0 0 0 0 0 1 0 0 0 0 0 0 
-0 0 1 0 0 1 0 0 0 0 0 0 
-
